Props stands for properties and are immutable and read-only properties.
With the help or props we can make the component reusable and flexible.

We can pass the string, number, boolean, object and array in props...

Like in the JS... Components are functions and arguments or parameters passed inside the function are props in react.

We can pass the default props and also we can destructure the props using the curly braces... With the help of destructuring....