props drilling
It is a process of passing data from parent component to deeply nexted child component via intermediates components that do not neccessarily need the data themeselves.

It is issue while passing the data from top level parent to deep level child component and intermediate component do not need the props (data) they are just carrying the data but do not use the props.

Solution of the props drilling is context API.